Moved @tryghost/oembed-service to services folder
fix https://linear.app/ghost/issue/ENG-2040/move-tryghostoembed-service-to-services-folder

- this moves the oembed service code to the services folder so we keep
  everything together
- removes the package as it's no longer needed
