+++ compose.yml
@@ -95,7 +95,6 @@ services:
       - node_modules_ghost_mw-version-match:/home/ghost/ghost/mw-version-match/node_modules:delegated
       - node_modules_ghost_mw-vhost:/home/ghost/ghost/mw-vhost/node_modules:delegated
       - node_modules_ghost_nql-filter-expansions:/home/ghost/ghost/nql-filter-expansions/node_modules:delegated
-      - node_modules_ghost_oembed-service:/home/ghost/ghost/oembed-service/node_modules:delegated
       - node_modules_ghost_offers:/home/ghost/ghost/offers/node_modules:delegated
       - node_modules_ghost_package-json:/home/ghost/ghost/package-json/node_modules:delegated
       - node_modules_ghost_payments:/home/ghost/ghost/payments/node_modules:delegated
@@ -272,7 +271,6 @@ volumes:
   node_modules_ghost_mw-version-match: {}
   node_modules_ghost_mw-vhost: {}
   node_modules_ghost_nql-filter-expansions: {}
-  node_modules_ghost_oembed-service: {}
   node_modules_ghost_offers: {}
   node_modules_ghost_package-json: {}
   node_modules_ghost_payments: {}
@@ -290,7 +288,6 @@ volumes:
   node_modules_ghost_stats-service: {}
   node_modules_ghost_stripe: {}
   node_modules_ghost_tiers: {}
-  node_modules_ghost_update-check-service: {}
   node_modules_ghost_verification-trigger: {}
   node_modules_ghost_version-notifications-data-service: {}
   node_modules_ghost_webmentions: {}
