+++ ghost/core/core/server/services/oembed/service.js
@@ -2,8 +2,8 @@ const config = require('../../../shared/config');
 const storage = require('../../adapters/storage');
 const externalRequest = require('../../lib/request-external');
 
-const OEmbed = require('@tryghost/oembed-service');
-const oembed = new OEmbed({config, externalRequest, storage});
+const OEmbedService = require('./OEmbedService');
+const oembed = new OEmbedService({config, externalRequest, storage});
 
 const NFT = require('./NFTOEmbedProvider');
 const nft = new NFT({
