+++ ghost/core/package.json
@@ -139,7 +139,6 @@
     "@tryghost/mw-vhost": "0.0.0",
     "@tryghost/nodemailer": "0.3.45",
     "@tryghost/nql": "0.12.7",
-    "@tryghost/oembed-service": "0.0.0",
     "@tryghost/package-json": "0.0.0",
     "@tryghost/post-revisions": "0.0.0",
     "@tryghost/posts-service": "0.0.0",
@@ -171,6 +170,7 @@
     "brute-knex": "4.0.1",
     "bson-objectid": "2.0.4",
     "chalk": "4.1.2",
+    "charset": "1.0.1",
     "cheerio": "0.22.0",
     "common-tags": "1.8.2",
     "compression": "1.8.0",
@@ -193,6 +193,7 @@
     "got": "11.8.6",
     "gscan": "4.47.0",
     "human-number": "2.0.4",
+    "iconv-lite": "0.6.3",
     "image-size": "1.2.0",
     "intl": "1.2.5",
     "intl-messageformat": "5.4.3",
@@ -205,6 +206,15 @@
     "knex-migrator": "5.3.0",
     "lodash": "4.17.21",
     "luxon": "3.5.0",
+    "metascraper": "5.45.15",
+    "metascraper-author": "5.45.10",
+    "metascraper-description": "5.45.10",
+    "metascraper-image": "5.45.10",
+    "metascraper-logo": "5.45.10",
+    "metascraper-logo-favicon": "5.42.0",
+    "metascraper-publisher": "5.45.10",
+    "metascraper-title": "5.45.10",
+    "metascraper-url": "5.45.10",
     "moment": "2.24.0",
     "moment-timezone": "0.5.45",
     "multer": "1.4.4",
@@ -219,6 +229,7 @@
     "stoppable": "1.1.0",
     "superagent": "5.1.0",
     "superagent-throttle": "1.0.1",
+    "tough-cookie": "4.1.4",
     "xml": "1.0.1"
   },
   "optionalDependencies": {
