--- ghost/oembed-service/package.json
@@ -1,44 +1,0 @@
-{
-  "name": "@tryghost/oembed-service",
-  "version": "0.0.0",
-  "repository": "https://github.com/TryGhost/Ghost/tree/main/packages/oembed-service",
-  "author": "Ghost Foundation",
-  "private": true,
-  "main": "index.js",
-  "scripts": {
-    "dev": "echo \"Implement me!\"",
-    "test:unit": "NODE_ENV=testing c8 --all --reporter text --reporter cobertura -- mocha --reporter dot './test/**/*.test.js'",
-    "test": "yarn test:unit",
-    "lint:code": "eslint *.js lib/ --ext .js --cache",
-    "lint": "yarn lint:code && yarn lint:test",
-    "lint:test": "eslint -c test/.eslintrc.js test/ --ext .js --cache"
-  },
-  "files": [
-    "index.js",
-    "lib"
-  ],
-  "devDependencies": {
-    "c8": "8.0.1",
-    "mocha": "10.2.0"
-  },
-  "dependencies": {
-    "@extractus/oembed-extractor": "3.2.1",
-    "@tryghost/errors": "1.3.5",
-    "@tryghost/logging": "2.4.21",
-    "@tryghost/tpl": "0.1.32",
-    "charset": "1.0.1",
-    "cheerio": "0.22.0",
-    "iconv-lite": "0.6.3",
-    "lodash": "4.17.21",
-    "metascraper": "5.45.15",
-    "metascraper-author": "5.45.10",
-    "metascraper-description": "5.45.10",
-    "metascraper-image": "5.45.10",
-    "metascraper-logo": "5.45.10",
-    "metascraper-logo-favicon": "5.42.0",
-    "metascraper-publisher": "5.45.10",
-    "metascraper-title": "5.45.10",
-    "metascraper-url": "5.45.10",
-    "tough-cookie": "4.1.4"
-  }
-}
